AI_Projects/Computer_Vision/Face_Recognition/requirements.txt
docker-compose
numpy
opencv-python
pandas
pymilvus
python-dotenv
%torch
%torchvision
ultralytics